core-*, stress-*: use %jd and (intmax_t) cast for all pid_t formatting

+++ core-helper.c
@@ -2798,8 +2798,8 @@ int stress_drop_capabilities(const char *name)
 
 	ret = capget(&uch, ucd);
 	if (ret < 0) {
-		pr_fail("%s: capget on PID %d failed: errno=%d (%s)\n",
-			name, uch.pid, errno, strerror(errno));
+		pr_fail("%s: capget on PID %jd failed: errno=%d (%s)\n",
+			name, (intmax_t)uch.pid, errno, strerror(errno));
 		return -1;
 	}
 
@@ -2819,8 +2819,8 @@ int stress_drop_capabilities(const char *name)
 
 	ret = capset(&uch, ucd);
 	if (ret < 0) {
-		pr_fail("%s: capset on PID %d failed: errno=%d (%s)\n",
-			name, uch.pid, errno, strerror(errno));
+		pr_fail("%s: capset on PID %jd failed: errno=%d (%s)\n",
+			name, (intmax_t)uch.pid, errno, strerror(errno));
 		return -1;
 	}
 #if defined(HAVE_PRCTL) &&		\
@@ -2830,9 +2830,9 @@ int stress_drop_capabilities(const char *name)
 	if (ret < 0) {
 		/* Older kernels that don't support this prctl throw EINVAL */
 		if (errno != EINVAL) {
-			pr_inf("%s: prctl PR_SET_NO_NEW_PRIVS on PID %d failed: "
+			pr_inf("%s: prctl PR_SET_NO_NEW_PRIVS on PID %jd failed: "
 				"errno=%d (%s)\n",
-				name, uch.pid, errno, strerror(errno));
+				name, (intmax_t)uch.pid, errno, strerror(errno));
 		}
 		return -1;
 	}
